Allow template to be given as default 1st parameter
Make it possible to lint with giving just the path to the template
(positional argument). The first parameter is now handled as the
template:

`cfn-lint path/to/file.yaml` followed by optional parameters: `cfn-lint
path/to.file.yaml -i W3010`

This is done to simplify the usage to the linter. The `—template` (or
`-t`) option is still available and will be used if no positional
argument is specified. This is for backwards compatibility and full
freedom in usage