Fix an incompatibility with samtranslator 1.10.0 (#702)
`cfn-python-lint` suppressed some logging from `samtranslator`. As it
turned out, the logging itself got already removed in `samtranslator`
1.7.0, while they retained the import of the logging module. This import
got finally removed with `samtranslator` 1.10.0, which breaks
`cfn-python-lint` when using version 1.10.0 of `samtranslator`.

This commit removes the suppression of the logging in `cfn-python-lint`,
as no logging happens in the supported `samtranslator` versions anymore
anyway.