Validate CertificateManager::Certificate::DomainValidationOptions
This introduces validation of CertificateManager::Certificate::DomainValidationOptions to ensure that the ValidationDomain is a superdomain of the domain being validated.

Eg: "subdomain.example.com" can be validated by itself or "example.com", but not by "example.net".

Deployment of the stack fails if this condition is not met, so I've introduced this as an error-level rule.