refactor: move list of stateful resource types to data file (#1897)
The CDK would like to use the same data that cfn-lint is using to
help our code generation.

This change moves the list of stateful resource types to a JSON file,
which is easier to parse for the CDK build process than a Python source
file. This will help keep the source of truth in sync between the
two projects.

This is not supposed to be the be-all-and-end-all solution of
integration between the two projects. I'm aware it's quite ad-hoc;
but let's try something low-friction and see how it goes?