Fix rules for Iops property and EBS volume types.
The 'Iops' property on an EBS volume is required for io1 and io2
volumes, optional for gp3 volumes, and not supported for gp2, st1, sc1
and standard volumes.

The previous code treated the property as required for io1 and not
supported for all other volume types.  This gave the wrong outcome for
io2 and gp3 specifically, and is generally the wrong move for any new
volume types that are added in the future.

Fixes #1977.