:bug: Add a check for comments and new lines in custom rules
In the documentation, comments are allowed to start with `#`.
Currently, cfn-lint throws an error of `not in supported operators` which means the comment is being parsed as a custom error.
This is now fixed to allow any amount of comments anywhere in the custom rules file.

New lines are now allowed just for readability.
Currently, cfn-lint throws a traceback error of `UnboundLocalError: cannot access local variable 'operator' where it is not associated with a value` which is not an intuitive error message just for a blank new line.
This is now fixed to allow new lines anywhere in the custom rules file.